feat: implement DNSCrypt client and VPN service
Integrate DNSCrypt logic and `VpnService` for encrypted traffic handling. Additionally, downgrade Room dependencies, update KSP versions, refactor UI imports, and optimize Gradle configuration for stability.

+++ app/src/main/java/com/example/ui/screens/DashboardScreen.kt
@@ -186,7 +186,7 @@ fun DashboardScreen(
         // 35% AI Security CPU Allocation Engine Card
         item {
             AiCpuAllocationCard(
-                aiAllocation = aiAllocation,
+                aiAllocation = aiAllocation.toFloat(),
                 aiStatus = aiStatus,
                 isProtected = isProtected,
                 bytesIn = bytesIn,
